Added thread controller so program can start at different points.
