+++ OSCLeash.py
@@ -1,202 +1,58 @@
-from asyncore import dispatcher
-from threading import Lock,Thread,Timer
-from pythonosc.osc_server import BlockingOSCUDPServer
-from pythonosc.dispatcher import Dispatcher
 from pythonosc.udp_client import SimpleUDPClient
-from dataclasses import dataclass
+from threading import Thread
 import json
 import os
-import sys
-import ctypes
-import time
 
-# Set window name on the Window
-if os.name == 'nt':
-    ctypes.windll.kernel32.SetConsoleTitleW("OSCLeash")
+from Controllers.DataController import DefaultConfig, ConfigSettings, Leash
+from Controllers.PackageController import Package
+from Controllers.ThreadController import Program
 
-# Console Clear
-def cls():
-    """Clears Console"""
-    os.system('cls' if os.name == 'nt' else 'clear')
-
-# Source Path
-def resource_path(relative_path):
-    """Gets absolute path from relative path"""
-    base_path = getattr(sys, '_MEIPASS', os.path.dirname(os.path.abspath(__file__)))
-    return os.path.join(base_path, relative_path)
-
-#Load Config
-try:
-    config = json.load(open(os.path.join(os.path.join(resource_path('config.json')))))
-    IP = config["IP"]
-    ListeningPort = config["ListeningPort"]
-    SendingPort = config["SendingPort"]
-    RunDeadzone = config["RunDeadzone"]
-    WalkDeadzone = config["WalkDeadzone"]
-    StrengthMultiplier = config["StrengthMultiplier"]
-    ActiveDelay = config["ActiveDelay"]
-    InactiveDelay = config["InactiveDelay"]
-    Logging = config["Logging"]
-    XboxJoystickMovement = config["XboxJoystickMovement"]
-    cls()
-    print("Successfully read config.")
-except Exception as e: 
-    cls()
-    print('\x1b[1;31;40m' + 'Missing or incorrect config file. Loading default values.  ' + '\x1b[0m')
-    print(e,"was the exception")
-    IP = "127.0.0.1"
-    ListeningPort = 9001
-    SendingPort = 9000
-    RunDeadzone = 0.70
-    WalkDeadzone = 0.15
-    StrengthMultiplier = 1.2    
-    ActiveDelay = .1
-    InactiveDelay = .5
-    Logging = True
-    XboxJoystickMovement = False
-
-# Settings confirmation
-print('\x1b[1;32;40m' + 'OSCLeash is Running!' + '\x1b[0m')
-if IP == "127.0.0.1":
-    print("IP: Localhost")
-else:  
-    print("IP: Not Localhost? Wack.")
-print("Listening on port", ListeningPort)
-print("Sending on port",SendingPort)
-print("Running Deadzone of {:.0f}".format(RunDeadzone*100)+"% stretch")
-print("Walking Deadzone of {:.0f}".format(WalkDeadzone*100)+"% stretch")
-print("Delays of {:.0f}".format(ActiveDelay*1000),"& {:.0f}".format(InactiveDelay*1000),"ms")
-#print("Inactive delay of {:.0f}".format(InactiveDelay*1000),"ms")
-if XboxJoystickMovement:
+def createDefaultConfigFile(configPath): # Creates a default config
     try:
-        import vgamepad as vg
-        gamepad = vg.VX360Gamepad()
-        print("Emulating Xbox 360 Controller for input instead of OSC")
+        with open("config.json", "w") as cf:
+            json.dump(DefaultConfig, cf)
+
+        print("Default config file created")
+
     except Exception as e:
         print(e)
-        print('\x1b[1;31;40m' + 'Tool required for controller emulation not installed. Check the docs.' + '\x1b[0m') 
+        exit()
 
-@dataclass
-class LeashParameters:
-    LeashWasGrabbed: bool = False
-    LeashGrabbed: bool = False
-    LeashStretch: float = 0
-    Z_Positive: float = 0
-    Z_Negative: float = 0
-    X_Positive: float = 0
-    X_Negative: float = 0
+if __name__ == "__main__":
 
-leash = LeashParameters()
+    #*************Setup*************#
+    program = Program()
+    program.setWindowTitle()
+    program.cls()
 
-statelock = Lock()
-dispatcher = Dispatcher()
-
-#Recieve paramaters and adjust values
-def OnRecieve(address,value):
-    parameter = address.split("/")[3]
-    statelock.acquire()
-    match parameter:
-        case "Leash_IsGrabbed":
-            leash.LeashGrabbed=value
-        case "Leash_Stretch":
-            leash.LeashStretch=value
-        case "Leash_Z+":
-            leash.Z_Positive=value
-        case "Leash_Z-":
-            leash.Z_Negative=value
-        case "Leash_X+":
-            leash.X_Positive=value
-        case "Leash_X-":
-            leash.X_Negative=value          
-    #print(f"{parameter}: {value}") #This Prints every input
-    statelock.release()
-
-# Paramaters to read
-dispatcher.map("/avatar/parameters/Leash_Z+",OnRecieve) #Z Positive
-dispatcher.map("/avatar/parameters/Leash_Z-",OnRecieve) #Z Negative
-dispatcher.map("/avatar/parameters/Leash_X+",OnRecieve) #X Positive
-dispatcher.map("/avatar/parameters/Leash_X-",OnRecieve) #X Negative
-dispatcher.map("/avatar/parameters/Leash_Stretch",OnRecieve) #Physbone Stretch Value
-dispatcher.map("/avatar/parameters/Leash_IsGrabbed",OnRecieve) #Physbone Grab Status
-#dispatcher.set_default_handler(OnRecieve) #This recieves everything, I think?
-
-# Set up UDP OSC client   
-oscClient = SimpleUDPClient(IP, SendingPort) 
-def StartServer():
-    try:
-        print("server")
-        server = BlockingOSCUDPServer((IP,ListeningPort),dispatcher)
-        server.serve_forever()
-    except:
-        print('\x1b[1;31;41m' + '                                                            ' + '\x1b[0m')
-        print('\x1b[1;31;40m' + '  Warning: An application is already running on this port!  ' + '\x1b[0m')
-        print('\x1b[1;31;41m' + '                                                            ' + '\x1b[0m')
-
-#clamp float values between -1 and 1
-def clamp (n):
-    return max(-1.0, min(n, 1.0))
-
-# Run Leash
-def LeashRun():
-    statelock.acquire()
-
-    #Maths 
-    VerticalOutput = clamp((leash.Z_Positive-leash.Z_Negative) * leash.LeashStretch * StrengthMultiplier)
-    HorizontalOutput = clamp((leash.X_Positive-leash.X_Negative) * leash.LeashStretch * StrengthMultiplier)
-
-    #Read Grab state
-    LeashGrabbed = leash.LeashGrabbed
-
-    if LeashGrabbed == True: #Leash is grabbed
-        leash.LeashWasGrabbed = True #Has been grabbed 
-
-    LeashReleased = leash.LeashGrabbed != leash.LeashWasGrabbed #Do Leash Grab states line up?
-    if LeashReleased == True: #Reset Leash grab state
-        leash.LeashWasGrabbed = False
-    
-    statelock.release()
-
-    if LeashGrabbed == True: #GrabCheck
-        if leash.LeashStretch > RunDeadzone: #Running deadzone
-            LeashOutput(VerticalOutput,HorizontalOutput,1)
-        elif leash.LeashStretch > WalkDeadzone: #Walking deadzone
-            LeashOutput(VerticalOutput,HorizontalOutput,0)
-        else: #Not stretched enough to move.
-            LeashOutput(0.0,0.0,0)
-    elif LeashReleased == True: #Leash was Dropped, send stop movement.
-        LeashReleased = False
-        LeashOutput(0.0,0.0,0)
-        time.sleep(InactiveDelay)
-        LeashOutput(0.0,0.0,0) #Double output, just in case.
+    # Test if Config file exists. Create the default if it does not.
+    configRelativePath = "./config.json"
+    if not os.path.exists(configRelativePath):
+        print("Config file was not found...", "\nCreating default config file...")
+        createDefaultConfigFile(configRelativePath)
     else:
-        time.sleep(InactiveDelay) #Add 500ms when not grabbed; save on performance?
+        print("Config file found\n")
 
-    #Wait 100ms, edge of human reaction time, probably?
-    Timer(ActiveDelay,LeashRun).start()
+    configData = json.load(open(configRelativePath)) # Config file should be prepared at this point.
+    settings = ConfigSettings(configData) # Get settings from config file
+    
+    # Collect Data for leash
+    
+    # TODO: Only currently works for one leash (Prepped for more)
+    # Notes: There is only one source of contacts for all Leashes. This means
+    #   that the pull result is dependent on the FIRST leash pulled. In a 
+    #   case that multiple are pulled, pick one that is in control.
+    
+    leashes = [] 
+    for leashName in configData["PhysboneParameters"]:
+        leashes.append(Leash(leashName, configData["DirectionalParamaters"], settings))
 
-# Output OSC
-def LeashOutput(VerticalOutput:float,HorizontalOutput:float,RunOutput):
-    if XboxJoystickMovement: #Xbox Emulation: REMOVE LATER WHEN OSC IS FIXED
-        gamepad.left_joystick_float(x_value_float=HorizontalOutput, y_value_float=VerticalOutput)
-        if RunOutput == 1:
-            gamepad.press_button(button=vg.XUSB_BUTTON.XUSB_GAMEPAD_LEFT_SHOULDER)      
-        else:
-            gamepad.release_button(button=vg.XUSB_BUTTON.XUSB_GAMEPAD_LEFT_SHOULDER)
-        gamepad.update()
-    else: #Normal function
-        oscClient.send_message("/input/Vertical", VerticalOutput)
-        oscClient.send_message("/input/Horizontal", HorizontalOutput)
-        oscClient.send_message("/input/Run", RunOutput)
+    # Manage data coming in
+    package = Package()
 
-    if Logging:
-        print(
-        "{:.2f}".format(VerticalOutput),"&",
-        "{:.2f}".format(HorizontalOutput),"&",
-        RunOutput)
-        #print("",VerticalOutput,"&",HorizontalOutput,"&",RunOutput) #Way too many decimale places debug
+    for leash in leashes:
+        package.listen(leash)
 
-thread=Thread(target=StartServer)
-thread.start()
-print("test")
-LeashRun()
-LeashOutput(0.0,0.0,0)
+    Thread(target=package.runServer, args=(settings.IP, settings.ListeningPort)).start()
+    Thread(target=program.leashRun, args=(leashes[0],)).start()
+
